Refactor Trigger Input Dialog into modular components
