+++ internal-packages/workflow-designer-ui/src/header/ui/trigger-input-dialog/dialog.tsx
@@ -1,402 +1,39 @@
 import {
 	type ActionNode,
-	type FlowTrigger,
 	type Generation,
 	type Input,
-	type ParameterItem,
 	type TriggerNode,
-	type WorkspaceId,
 	isActionNode,
 } from "@giselle-sdk/data-type";
-import type { githubTriggers } from "@giselle-sdk/flow";
 import { useGenerationRunnerSystem } from "@giselle-sdk/giselle-engine/react";
 import { buildWorkflowFromNode } from "@giselle-sdk/workflow-utils";
 import { clsx } from "clsx/lite";
 import { useWorkflowDesigner } from "giselle-sdk/react";
-import { AlertTriangleIcon, LoaderIcon, PlayIcon, XIcon } from "lucide-react";
+import { AlertTriangleIcon, PlayIcon, XIcon } from "lucide-react";
 import { Dialog } from "radix-ui";
 import {
-	type ButtonHTMLAttributes,
 	type FormEventHandler,
-	type ReactNode,
 	useCallback,
 	useMemo,
 	useRef,
 	useState,
 } from "react";
-import type { z } from "zod";
-import { useTrigger } from "../../hooks/use-trigger";
-import { useToasts } from "../../ui/toast";
 
-export function Button({
-	leftIcon: LeftIcon,
-	rightIcon: RightIcon,
-	loading = false,
-	disabled = false,
-	children,
-	...props
-}: {
-	leftIcon?: ReactNode;
-	rightIcon?: ReactNode;
-	loading?: boolean;
-	disabled?: boolean;
-} & Omit<ButtonHTMLAttributes<HTMLButtonElement>, "disabled">) {
-	const isDisabled = loading || disabled;
-
-	return (
-		<button
-			type="button"
-			className={clsx(
-				"bg-white-900 px-[8px] rounded-[4px] py-[4px] text-[14px] flex items-center gap-[4px] outline-none text-black-900",
-				"data-[loading=true]:cursor-not-allowed data-[loading=true]:opacity-60",
-				"data-[disabled=true]:cursor-not-allowed data-[disabled=true]:opacity-60",
-				"data-[loading=false]:data-[disabled=false]:cursor-pointer",
-			)}
-			data-loading={loading}
-			data-disabled={isDisabled}
-			disabled={isDisabled}
-			{...props}
-		>
-			{loading ? <LoaderIcon className="size-[14px] animate-spin" /> : LeftIcon}
-			<div>{children}</div>
-			{!loading && RightIcon}
-		</button>
-	);
-}
-
-export function buttonLabel(node: TriggerNode) {
-	switch (node.content.provider) {
-		case "manual":
-			return "Start Manual Flow";
-		case "github":
-			return "Test with dummy data";
-		default: {
-			const _exhaustiveCheck: never = node.content.provider;
-			throw new Error(`Unhandled trigger provider type: ${_exhaustiveCheck}`);
-		}
-	}
-}
-
-interface FormInput {
-	name: string;
-	label: string;
-	type: "text" | "multiline-text" | "number";
-	required: boolean;
-}
-
-type GithubEventInputMap = {
-	[K in keyof typeof githubTriggers]: {
-		[K2 in keyof z.infer<
-			(typeof githubTriggers)[K]["event"]["payloads"]
-		>]: Omit<FormInput, "name">;
-	};
-};
-
-// Define the input fields for each GitHub event type
-const githubEventInputs: GithubEventInputMap = {
-	"github.issue.created": {
-		issueNumber: {
-			label: "Issue Number",
-			type: "number",
-			required: true,
-		},
-		title: {
-			label: "Title",
-			type: "text",
-			required: true,
-		},
-		body: {
-			label: "Body",
-			type: "multiline-text",
-			required: false,
-		},
-	},
-	"github.issue.closed": {
-		issueNumber: {
-			label: "Issue Number",
-			type: "number",
-			required: true,
-		},
-		title: {
-			label: "Title",
-			type: "text",
-			required: true,
-		},
-		body: {
-			label: "Body",
-			type: "multiline-text",
-			required: false,
-		},
-	},
-	"github.issue_comment.created": {
-		issueNumber: {
-			label: "Issue Number",
-			type: "number",
-			required: true,
-		},
-		issueTitle: {
-			label: "Issue Title",
-			type: "text",
-			required: true,
-		},
-		issueBody: {
-			label: "Issue Body",
-			type: "multiline-text",
-			required: true,
-		},
-		body: {
-			label: "Issue Comment",
-			type: "multiline-text",
-			required: true,
-		},
-	},
-	"github.pull_request_comment.created": {
-		issueNumber: {
-			label: "Pull Request Number",
-			type: "number",
-			required: true,
-		},
-		issueTitle: {
-			label: "Pull Request Title",
-			type: "text",
-			required: true,
-		},
-		issueBody: {
-			label: "Pull Request Body",
-			type: "multiline-text",
-			required: true,
-		},
-		body: {
-			label: "Pull Request Comment",
-			type: "multiline-text",
-			required: true,
-		},
-	},
-	"github.pull_request.ready_for_review": {
-		title: {
-			label: "Title",
-			type: "text",
-			required: true,
-		},
-		body: {
-			label: "Body",
-			type: "multiline-text",
-			required: false,
-		},
-		number: {
-			label: "Number",
-			type: "number",
-			required: true,
-		},
-		diff: {
-			label: "diff",
-			type: "multiline-text",
-			required: false,
-		},
-		pullRequestUrl: {
-			label: "Pull request URL",
-			type: "text",
-			required: true,
-		},
-	},
-	"github.pull_request.closed": {
-		title: {
-			label: "Title",
-			type: "text",
-			required: true,
-		},
-		body: {
-			label: "Body",
-			type: "multiline-text",
-			required: false,
-		},
-		number: {
-			label: "Number",
-			type: "number",
-			required: true,
-		},
-		pullRequestUrl: {
-			label: "Pull request URL",
-			type: "text",
-			required: true,
-		},
-	},
-	"github.pull_request.opened": {
-		title: {
-			label: "Title",
-			type: "text",
-			required: true,
-		},
-		body: {
-			label: "Body",
-			type: "multiline-text",
-			required: false,
-		},
-		number: {
-			label: "Number",
-			type: "number",
-			required: true,
-		},
-		diff: {
-			label: "diff",
-			type: "multiline-text",
-			required: false,
-		},
-		pullRequestUrl: {
-			label: "Pull request URL",
-			type: "text",
-			required: true,
-		},
-	},
-};
-
-function createInputsFromTrigger(
-	trigger: FlowTrigger | undefined,
-): FormInput[] {
-	if (trigger === undefined) {
-		return [];
-	}
-
-	switch (trigger.configuration.provider) {
-		case "github": {
-			const inputDefs = githubEventInputs[trigger.configuration.event.id];
-			return Object.entries(inputDefs).map(([name, def]) => ({
-				name,
-				label: def.label,
-				type: def.type,
-				required: def.required,
-			}));
-		}
-		case "manual": {
-			return trigger.configuration.event.parameters.map((parameter) => ({
-				name: parameter.id,
-				label: parameter.name,
-				type: parameter.type,
-				required: parameter.required,
-			}));
-		}
-		default: {
-			const _exhaustiveCheck: never = trigger.configuration;
-			throw new Error(`Unhandled provider: ${_exhaustiveCheck}`);
-		}
-	}
-}
-function parseFormInputs(inputs: FormInput[], formData: FormData) {
-	const errors: Record<string, string> = {};
-	const values: Record<string, string | number> = {};
-
-	for (const input of inputs) {
-		const formDataEntryValue = formData.get(input.name);
-		const value = formDataEntryValue
-			? formDataEntryValue.toString().trim()
-			: "";
-
-		if (input.required && value === "") {
-			errors[input.name] = `${input.label} is required`;
-			continue;
-		}
-
-		if (value === "") {
-			values[input.name] = "";
-			continue;
-		}
-
-		switch (input.type) {
-			case "text":
-			case "multiline-text":
-				values[input.name] = value;
-				break;
-			case "number": {
-				const numValue = Number(value);
-				if (Number.isNaN(numValue)) {
-					errors[input.name] = `${input.label} must be a valid number`;
-				} else {
-					values[input.name] = numValue;
-				}
-				break;
-			}
-			default: {
-				const _exhaustiveCheck: never = input.type;
-				throw new Error(`Unhandled input type: ${_exhaustiveCheck}`);
-			}
-		}
-	}
-
-	return { errors, values };
-}
-
-function toParameterItems(
-	inputs: FormInput[],
-	values: Record<string, string | number>,
-): ParameterItem[] {
-	const items: ParameterItem[] = [];
-	for (const input of inputs) {
-		const value = values[input.name];
-		if (value === undefined || value === "") {
-			continue;
-		}
-		switch (input.type) {
-			case "text":
-			case "multiline-text":
-				items.push({
-					type: "string",
-					name: input.name,
-					value: value as string,
-				});
-				break;
-			case "number":
-				items.push({
-					type: "number",
-					name: input.name,
-					value: value as number,
-				});
-				break;
-			default: {
-				const _exhaustiveCheck: never = input.type;
-				throw new Error(`Unhandled input type: ${_exhaustiveCheck}`);
-			}
-		}
-	}
-	return items;
-}
-
-function createGenerationsForFlow(
-	flow: NonNullable<ReturnType<typeof buildWorkflowFromNode>>,
-	inputs: FormInput[],
-	values: Record<string, string | number>,
-	createGeneration: ReturnType<
-		typeof useGenerationRunnerSystem
-	>["createGeneration"],
-	workspaceId: WorkspaceId,
-) {
-	const generations: Generation[] = [];
-	for (const job of flow.jobs) {
-		for (const operation of job.operations) {
-			const parameterItems =
-				operation.node.content.type === "trigger"
-					? toParameterItems(inputs, values)
-					: [];
-			const generation = createGeneration({
-				origin: { type: "workspace", id: workspaceId },
-				inputs:
-					parameterItems.length > 0
-						? [{ type: "parameters", items: parameterItems }]
-						: [],
-				...operation.generationTemplate,
-			});
-			generations.push(generation);
-		}
-	}
-	return generations;
-}
+import { useTrigger } from "../../../hooks/use-trigger";
+import { useToasts } from "../../../ui/toast";
+import { Button } from "./button";
+import {
+	type FormInput,
+	buttonLabel,
+	createInputsFromTrigger,
+	parseFormInputs,
+} from "./helpers";
+import { createGenerationsForFlow } from "./workflow";
 
 export function TriggerInputDialog({
 	node,
 	onClose,
-}: {
-	node: TriggerNode;
-	onClose: () => void;
-}) {
+}: { node: TriggerNode; onClose: () => void }) {
 	const { data: trigger, isLoading } = useTrigger(node);
 	const [validationErrors, setValidationErrors] = useState<
 		Record<string, string>
@@ -424,9 +61,9 @@ export function TriggerInputDialog({
 			flow === null
 				? []
 				: flow.nodes
-						.filter((node) => isActionNode(node, "github"))
-						.map((node) => {
-							const notConnectedRequiredInputs = node.inputs.filter(
+						.filter((n) => isActionNode(n, "github"))
+						.map((n) => {
+							const notConnectedRequiredInputs = n.inputs.filter(
 								(input) =>
 									input.isRequired &&
 									!data.connections.some(
@@ -437,7 +74,7 @@ export function TriggerInputDialog({
 								return null;
 							}
 							return {
-								node,
+								node: n as ActionNode,
 								inputs: notConnectedRequiredInputs,
 							};
 						})
@@ -515,8 +152,8 @@ export function TriggerInputDialog({
 				await Promise.all(
 					job.operations.map(async (operation) => {
 						const generation = generations.find(
-							(generation) =>
-								generation.context.operationNode.id ===
+							(g) =>
+								g.context.operationNode.id ===
 								operation.generationTemplate.operationNode.id,
 						);
 						if (generation === undefined) {
@@ -529,7 +166,6 @@ export function TriggerInputDialog({
 					}),
 				);
 
-				// Close dialog after first job completes
 				if (jobIndex === 0) {
 					onClose();
 				}
@@ -550,6 +186,7 @@ export function TriggerInputDialog({
 	if (isLoading || trigger === undefined || flow === null) {
 		return null;
 	}
+
 	return (
 		<>
 			<div className="flex justify-between items-center mb-[14px]">
@@ -598,8 +235,8 @@ export function TriggerInputDialog({
 											>
 												<span className="font-medium">
 													{item.node.name || "Unnamed Action"}
-												</span>
-												{" - Missing: "}
+												</span>{" "}
+												- Missing:{" "}
 												{item.inputs.map((input) => input.label).join(", ")}
 											</li>
 										))}
