+++ internal-packages/workflow-designer-ui/src/header/ui/trigger-input-dialog/index.ts
@@ -1,0 +1,9 @@
+export { Button } from "./button";
+export {
+	buttonLabel,
+	type FormInput,
+	createInputsFromTrigger,
+	parseFormInputs,
+} from "./helpers";
+export { createGenerationsForFlow } from "./workflow";
+export { TriggerInputDialog } from "./dialog";
